Fix convert command to handle large numeric response values (#52)
The units output by default makes anything equal or larger than e8 display in scientific format. This is working as expected when significand is a float, but fails when it is an integer.
The fix converts the scientific notation output from units to what erlang expects - significand must be a float.