Remove CircleCI tests (#55)
Now that we are running all tests in GitHub workflows, there's no longer a need to run them in CircleCI too. Additionally, it seems the CircleCI convenience images are less frequently updated - for example, as of this writing the last supported Ubuntu base image is still 20.04 and Erlang/OTP 26.0 is not yet available (latest available is 25.3).